+++ src/lib.rs
@@ -21,4 +21,5 @@ pub mod routers;
 pub mod server;
 pub mod service_discovery;
 pub mod tokenizer;
+pub mod transport;
 pub mod tree;